Feature #334: Submit gLite and Dirac with specific user proxy

+++ src/main/java/fr/insalyon/creatis/gasw/output/DiracOutputUtil.java
@@ -43,6 +43,9 @@ import fr.insalyon.creatis.gasw.dao.JobDAO;
 import fr.insalyon.creatis.gasw.monitor.Monitor;
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 import org.apache.log4j.Logger;
 
 /**
@@ -57,30 +60,49 @@ public class DiracOutputUtil extends OutputUtil {
         super(startTime);
     }
 
-    public GaswOutput getOutputs(String jobID) {
+    public GaswOutput getOutputs(String jobID){
+        return getOutputs(jobID, "");
+    }
+    
+    public GaswOutput getOutputs(String jobID, String proxyFile) {
         try {
             JobDAO jobDAO = DAOFactory.getDAOFactory().getJobDAO();
             Job job = jobDAO.getJobByID(jobID);
 
             if (job.getStatus() != Monitor.Status.CANCELLED && job.getStatus() != Monitor.Status.STALLED) {
 
-                String exec = "dirac-wms-job-get-output " + jobID;
-                Process execution = Runtime.getRuntime().exec(exec);
-                execution.waitFor();
+                List exec = new ArrayList();
+                exec.add("dirac-wms-job-get-output");
+                exec.add(jobID);
+                ProcessBuilder builder = new ProcessBuilder(exec);
+                builder.redirectErrorStream(false);
+                Map<String, String> environment = builder.environment();
 
-                File stdOut = getStdFile(job, ".out", Constants.OUT_ROOT);
-                File stdErr = getStdFile(job, ".err", Constants.ERR_ROOT);
+                if (!proxyFile.isEmpty() && proxyFile != null){
+                    environment.put("X509_USER_PROXY", proxyFile);
+                }
 
-                File outTempDir = new File("./" + jobID);
-                outTempDir.delete();
+                Process process = builder.start();
 
-                int exitCode = parseStdOut(job, stdOut);
-                exitCode = parseStdErr(job, stdErr, exitCode);
+                process.waitFor();
 
-                File appStdOut = saveFile(job, ".app.out", Constants.OUT_ROOT, getAppStdOut());
-                File appStdErr = saveFile(job, ".app.err", Constants.ERR_ROOT, getAppStdErr());
+                if (process.exitValue() == 0) {
+                    File stdOut = getStdFile(job, ".out", Constants.OUT_ROOT);
+                    File stdErr = getStdFile(job, ".err", Constants.ERR_ROOT);
 
-                return new GaswOutput(jobID, exitCode, appStdOut, appStdErr, stdOut, stdErr);
+                    File outTempDir = new File("./" + jobID);
+                    outTempDir.delete();
+
+                    int exitCode = parseStdOut(job, stdOut);
+                    exitCode = parseStdErr(job, stdErr, exitCode);
+
+                    File appStdOut = saveFile(job, ".app.out", Constants.OUT_ROOT, getAppStdOut());
+                    File appStdErr = saveFile(job, ".app.err", Constants.ERR_ROOT, getAppStdErr());
+
+                    return new GaswOutput(jobID, exitCode, appStdOut, appStdErr, stdOut, stdErr);
+                } else{
+                   logger.info("Output files does not exist. Job was not executed! - jobID " + jobID);
+                }
 
             } else {
